Add a payments backup mechanism
When restoring a wallet, the user can now restore a backup
file if one is available to recover the payments history.

This is done manually, by browsing the device's file system
and selecting a file. The file is then decrypted, read, and
the payments database it contains are restored in the phoenix
internal db directory.

This step is optional and can be skipped. If there already
is a payments db in the  internal db directory, this step
is skipped automatically when restoring a wallet (will
typically happen because of the cloud backup mechanism
provided by google drive).